Revert "winebus: Do not touch autocenter on device init and device reset."
The removal of autocenter support from the SDL and udev input
backends was based on the belief that the device should not turn
auto centering on upon receiving a device reset command from the
front end as only some devices do it this way and others have
different ways of doing it.

This is a fundamental misunderstanding of what is going on. The
messages between the front end and the SDL and udev backends are
the front end telling the backends what to do. These are in no
way sent to the device, and have nothing to do with how autocenter
is turned on or off on the device. Rather the backends interpret
them and then makes SDL and udev input API calls to drive the
device, including turning on or off autocentering.

The reason the backends call their respective APIs to turn on
autocentering on receiving a reset is because that is they way
turn on autocenter is communicated to the backend. In the event
that the autocenter should not be on, the front end will follow
the reset with a stop all effects command which tells the backends
to call their underlying API to turn it back off.

If autocenter is different than a users experts it to be, either
* the program assumed a given initial state, didn't verify it or
  or set it, and the initial state was the other, or
* the backend API calls are failing to set it (the kernel driver
  either does not support setting it or is buggy).

I have written a test program that lets people test their devices
to ensure things are working correctly and discover what the initial
state is under Windows

https://github.com/twhitehead/issue-wine-ff-autocenter

A log file is generated for users to submit so initial states under
Windows can be discovered and broken drivers identified. This will
allow the unwanted autocenter issue of the author of the commit
being reverted to be properly figured out and addressed.

As a final note, the SDL and udev input backends are actually the
preferred backends for force feedback as they ultimately dispatch
to the kernel driver which can handle any required quirks as well
as all the different way autocentering can be implemented.

By comparison, the udev hidraw backend is quite limited. It
requires a PID compliant device without any quirks that uses
reset and stop all effect autocentering. Such devices are
perfectly supported by the kernel ff_pdiff driver, so they to
function perfectly with the SDL and udev input backends.

This reverts commit b0db6d50530023a19846f08f1c62f14803448f59.