dinput: Add ability to set initial autocenter via registry key
Programs should check and/or set DIPROP_AUTOCENTER to ensure it
is the value they want. Likely some do not though if the
developers never tested on devices where the Windows' default
wasn't correct for them.

These programs will be broken under Wine if the Wine default is
different than the Windows. This commit adds an 'Autocenter'
appkey (to the existing 'Joysticks' and 'MouseWarpOverride' keys)
so users can override the DIPROP_AUTOCENTER default in this case.

The Wine default is currently always DIPROPAUTOCENTER_ON, which
is the correct Windows default for the MS Sidewinder 2. I don't
know any other correct defaults as this is the only FF device I
have to test. I started a project to collect defaults (and verify
operations) for other devices though

https://github.com/twhitehead/issue-wine-ff-autocenter

so hopefully a good heuristic and a more comprehensive set of
defaults can be implemented at some point.