dinput: Add autocenter warning if aquired without check or set
Programs should check and/or set DIPROP_AUTOCENTER to ensure it
is the value they want. Likely some do not though if the
developers never tested on devices where the Windows' default
wasn't correct for them. These programs will be broken under
Wine if the Wine default is different than the Windows.

This commit logs a warning upon acquisition, if the program does
it without first checking or setting DIPROP_AUTOCENTER. The warning
tells the user what the default is, that it may be incorrect, and
that they can override it with the new 'Autocenter' app key.