dinput: Autocenter on default unless device name contains joystick
While ideally programs should check and turn autocenter off, it
is likely that many do not. The default autocenter under Wine has
been on. Tomasz Pakuta reported in !8744 that this is the wrong
default for wheels and it has been causing a problem for many in
the sim racing community.

This commit makes `DIPROPAUTOCENTER_OFF` the default value for
`DIPROP_AUTOCENTER` unless the device name contains "joystick".
Although this will miss some joysticks (those without joystick
in their name, like the "Logitech WingMan"), it should hopefully
be an overall win to have joysticks be the exception and not the
default as I read there are more wheels than joysticks.

Wrongly classified devices can still be overriden by the new user
app key "Autocenter". As in

[Software\\Wine\\DirectInput\\Autocenter]
"Device instance name"="on"

or

[Software\\Wine\\AppDefaults\\app.exe\\DirectInput\\Autocenter]
"Device instance name"="on"